Implement Issue #61

+++ API/app/Http/Controllers/Api/ShoppingListController.php
@@ -3,125 +3,30 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+// Our custom controller for checking permissions (read, update, delete) for lists
+// (some lists may be shared rather than owned, so this is needed for that)
+use App\Http\Controllers\ListPermissionsController;
+
 use App\Models\ShoppingList;
-use App\Models\SharedLink;
 use App\Models\SharedShoppingListPerm;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Str;
 use Illuminate\Support\Facades\DB;
 use App\Models\Route; // Import the Route model
 use Illuminate\Support\Facades\Schema; // Necessary for debugging the schema
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 define('DEBUG_MODE', 0);
 
 
 class ShoppingListController extends Controller
 {
+    use AuthorizesRequests;
+    
     public function index()
     {
         return response()->json(ShoppingList::all(), 200);
-    }
-
-    public function share($id, Request $request)
-    {
-        $shoppingList = ShoppingList::findOrFail($id);
-        Log::info("Shopping list id: " . $shoppingList->list_id . " and we started with id: " . $id);
-        //Log::info("shoppingList object: " . print_r($shoppingList, true));
-
-        $maxRetries = 10; // Maximum number of retries
-        $retryCount = 0;
-
-        do {
-            $token = (string) Str::uuid();
-            $exists = SharedLink::where('token', $token)->exists();
-            $retryCount++;
-        } while ($exists && $retryCount < $maxRetries);
-
-        if ($exists) {
-            // If a unique token could not be generated, handle the error
-            return response()->json(['error' => 'Unable to generate a unique link. Please try again later.'], 500);
-        }
-
-        // Extract permissions from request
-        $canUpdate = $request->can_update ?? false;
-        $canDelete = $request->can_delete ?? false;
-
-        Log::info("canUpdate: " . $canUpdate);
-        Log::info("canDelete: " . $canDelete);
-
-
-        // Store the link with the shopping list ID and permissions
-        SharedLink::create([
-            'shopping_list_id' => $shoppingList->list_id,
-            'token' => $token,
-            'expires_at' => now()->addDays(7), // Optional: set expiration date
-            'can_update' => $canUpdate,
-            'can_delete' => $canDelete,
-        ]);
-
-        // Create the full URL with the domain
-        $url = "https://www.speedcartapp.com/share/$token";
-
-        return response()->json(['link' => $url]);
-    }
-
-    public function verifyShare($token, Request $request)
-    {
-        Log::info("Received shopping-lists request from IP address " . $request->ip());
-
-        // Grab user from sanctum
-        $user = Auth::user();
-
-        // Retrieve the shared link
-        $sharedLink = SharedLink::where('token', $token)->first();
-
-        // Check if the link exists
-        if (!$sharedLink) {
-            abort(404, 'Link not found');
-        }
-        Log::info("LINK WAS FOUND!");
-
-        // Check if the link is expired
-        if (now()->greaterThan($sharedLink->expires_at)) {
-            Log::info("Link has expired");
-            // Remove the expired link from the database
-            SharedLink::where('token', $token)->delete();
-            abort(403, 'Link expired');
-        }
-
-        Log::info("Retrieving shopping list with id " . $sharedLink->shopping_list_id);
-
-        // Retrieve the shopping list
-        $shoppingList = ShoppingList::findOrFail($sharedLink->shopping_list_id);
-
-        Log::info("Got shopping list!");
-
-        // Here, you may want to update the permissions for the user or perform any other actions
-        // For instance, you might want to set permissions for the user, if applicable.
-        $currentUserId = $user->user_id; // Assuming the user is authenticated
-
-        if ($currentUserId) {
-            Log::info("Setting user permissions for user: " . $currentUserId);
-            // Create or update the permission record
-            SharedShoppingListPerm::updateOrCreate(
-                ['shopping_list_id' => $shoppingList->list_id, 'user_id' => $currentUserId],
-                [
-                    'shopping_list_id' => $shoppingList->list_id,
-                    'user_id' => $currentUserId,
-                    'can_update' => $sharedLink->can_update,
-                    'can_delete' => $sharedLink->can_delete,
-                ]
-            );
-        }
-        // Remove the used link from the database
-        SharedLink::where('token', $token)->delete();
-
-        // Retrieve the shopping list
-        $shoppingList = ShoppingList::findOrFail($sharedLink->shopping_list_id);
-
-        return response()->json($shoppingList, 201);
     }
 
     public function store(Request $request)
@@ -185,17 +90,9 @@ class ShoppingListController extends Controller
 
         $shoppingList = ShoppingList::findOrFail($id);
 
-        // Grab user from sanctum
-        $user = Auth::user();
+        // This will automatically call the `view` method in the ShoppingListPolicy
+        $this->authorize('view', $shoppingList); // Throws a 403 if not authorized
 
-        if (strcmp($shoppingList->user_id, $user->user_id)) {
-            // Check if they have permissions for this list in the permissions table (read permissions is implied if an entry is present; reading is the bare minimum)
-            if (SharedShoppingListPerm::where('shopping_list_id', $shoppingList->list_id)->where('user_id', $user->user_id)) {
-                return response()->json($shoppingList, 200);
-            } else {
-                return response()->json(["errorMessage" => "Unauthorized Request"], 403);
-            }
-        }
         return response()->json($shoppingList, 200);
     }
 
@@ -252,28 +149,8 @@ class ShoppingListController extends Controller
         // Find the shopping list by ID and update the name
         $shoppingList = ShoppingList::findOrFail($id);
 
-        // Grab user from sanctum
-        $user = Auth::user();
-
-        // NEW CODE HASN'T BEEN TESTED (see Git on left side in VSCode for untested blocks)
-        $sharedPermissionEntry = SharedShoppingListPerm::where('shopping_list_id', $shoppingList->list_id)
-            ->where('user_id', $user->user_id)
-            ->first(); // Retrieves the first matching entry or null if none found
-        // (THERE SHOULD BE AT MOST ONE ENTRY RETURNED, which is why ->first() should work)
-        
-        // Sharing will be a feature added here later
-        if (strcmp($shoppingList->user_id, $user->user_id)) {
-            // User isn't owner; check if user has delete permissions via share table
-            if ($sharedPermissionEntry) {
-                if ($sharedPermissionEntry->can_update) {
-                    // User has delete permission; delete the list title (items deleted via CASCADE deletion behavior)
-                    $shoppingList->delete();
-
-                    return response()->json(['message' => 'Shopping list deleted successfully'], 200);
-                }
-            }
-            return response()->json(["errorMessage" => "Unauthorized Request"], 403);
-        }
+        // This will automatically call the `update` method in the ShoppingListPolicy
+        $this->authorize('update', $shoppingList); // Throws a 403 if not authorized
         
         $shoppingList->name = $validatedData['name'];
         $shoppingList->save();
@@ -286,26 +163,9 @@ class ShoppingListController extends Controller
         try {
             $shoppingList = ShoppingList::findOrFail($id);
 
-            // Grab user from sanctum
-            $user = Auth::user();
-            $sharedPermissionEntry = SharedShoppingListPerm::where('shopping_list_id', $shoppingList->list_id)
-                ->where('user_id', $user->user_id)
-                ->first(); // Retrieves the first matching entry or null if none found
-            // (THERE SHOULD BE AT MOST ONE ENTRY RETURNED, which is why ->first() should work)
-
-            if (strcmp($shoppingList->user_id, $user->user_id)) {
-                // User isn't owner; check if user has delete permissions via share table
-                if ($sharedPermissionEntry) {
-                    if ($sharedPermissionEntry->can_delete) {
-                        // User has delete permission; delete the list title (items deleted via CASCADE deletion behavior)
-                        $shoppingList->delete();
-
-                        return response()->json(['message' => 'Shopping list deleted successfully'], 200);
-                    }
-                }
-                return response()->json(["errorMessage" => "Unauthorized Request"], 403);
-                //return response()->json(['message' => 'Shopping list deleted successfully'], 200);
-            }
+            // This will automatically call the `delete` method in the ShoppingListPolicy
+            $this->authorize('delete', $shoppingList); // Throws a 403 if not authorized
+        
             $shoppingList->delete();
 
             return response()->json(['message' => 'Shopping list deleted successfully'], 200);
